JBAI-13783. Extract a general interface for AIAgent
