+++ agents/agents-core/src/commonMain/kotlin/ai/grazie/code/agents/core/agent/AIAgentTool.kt
@@ -1,7 +1,7 @@
-package ai.grazie.code.agents.core
+package ai.grazie.code.agents.core.agent
 
-import ai.grazie.code.agents.core.AIAgentTool.AgentToolArgs
-import ai.grazie.code.agents.core.AIAgentTool.AgentToolResult
+import ai.grazie.code.agents.core.agent.AIAgentTool.AgentToolArgs
+import ai.grazie.code.agents.core.agent.AIAgentTool.AgentToolResult
 import ai.grazie.code.agents.core.model.agent.AIAgentConfig
 import ai.grazie.code.agents.core.model.agent.AIAgentStrategy
 import ai.grazie.code.agents.core.tools.*
@@ -9,20 +9,20 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 
-fun <TStrategy : AIAgentStrategy<TConfig>, TConfig : AIAgentConfig> AIAgent<TStrategy, TConfig>.asTool(
+fun <TStrategy : AIAgentStrategy<TConfig>, TConfig : AIAgentConfig> AIAgentBase<TStrategy, TConfig>.asTool(
     agentDescription: String,
     requestDescription: String = "Input for the task"
 ) = AIAgentTool<TStrategy, TConfig>(this, requestDescription, agentDescription)
 
 
 class AIAgentTool<TStrategy : AIAgentStrategy<TConfig>, TConfig : AIAgentConfig>(
-    val agent: AIAgent<TStrategy, TConfig>,
+    val agent: AIAgentBase<TStrategy, TConfig>,
     agentName: String,
     agentDescription: String,
     requestDescription: String = "Input for the task"
 ) : Tool<AgentToolArgs, AgentToolResult>() {
     @Serializable
-    data class AgentToolArgs(val request: String) : Tool.Args
+    data class AgentToolArgs(val request: String) : Args
 
     @Serializable
     data class AgentToolResult(
