+++ agents/agents-core/src/commonMain/kotlin/ai/grazie/code/agents/core/api/AIAgent.kt
@@ -1,0 +1,22 @@
+package ai.grazie.code.agents.core.api
+
+/**
+ * Represents a basic interface for AI agent.
+ */
+interface AIAgent {
+
+    /**
+     * Initiates the execution of the AI agent based on the provided input prompt.
+     *
+     * @param prompt the input text that guides the agent's behavior or response generation
+     */
+    suspend fun run(prompt: String)
+
+    /**
+     * Executes the AI agent with the given prompt and retrieves the resulting output.
+     *
+     * @param userPrompt the input text used to guide the behavior or response of the AI agent
+     * @return the response generated by the AI agent, or null if no result is available
+     */
+    suspend fun runAndGetResult(userPrompt: String): String?
+}
