+++ agents/agents-core/src/commonMain/kotlin/ai/grazie/code/agents/core/api/LocalAIAgentExt.kt
@@ -1,6 +1,6 @@
 package ai.grazie.code.agents.core.api
 
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.event.EventHandler
 import ai.grazie.code.agents.core.tools.ToolRegistry
