+++ agents/agents-core/src/commonMain/kotlin/ai/grazie/code/agents/core/feature/handler/AgentHandler.kt
@@ -1,7 +1,7 @@
 package ai.grazie.code.agents.core.feature.handler
 
 import ai.grazie.code.agents.core.annotation.InternalAgentsApi
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.entity.LocalAgentStrategy
 import ai.grazie.code.agents.core.agent.entity.stage.LocalAgentStage
 import ai.grazie.code.agents.core.environment.AgentEnvironment
