+++ agents/agents-core/src/jvmTest/kotlin/ai/grazie/code/agents/core/calculator/CalculatorTest.kt
@@ -2,7 +2,7 @@ package ai.grazie.code.agents.core.calculator
 
 import ai.grazie.code.agents.core.event.EventHandler
 import ai.grazie.code.agents.core.tools.ToolRegistry
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.agent.entity.LocalAgentStrategy
 import ai.grazie.code.agents.core.dsl.builder.forwardTo
