+++ agents/agents-examples/src/main/kotlin/ai/grazie/code/agents/example/calculator/Calculator.kt
@@ -1,6 +1,6 @@
 package ai.grazie.code.agents.example.calculator
 
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.api.AskUser
 import ai.grazie.code.agents.core.api.SayToUser
