+++ agents/agents-examples/src/main/kotlin/ai/grazie/code/agents/example/memory/ProjectAnalyzerAgent.kt
@@ -1,6 +1,6 @@
 package ai.grazie.code.agents.example.memory
 
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.dsl.builder.forwardTo
 import ai.grazie.code.agents.core.dsl.builder.strategy
