+++ agents/agents-examples/src/main/kotlin/ai/grazie/code/agents/example/planner/PlannerAgentExample.kt
@@ -1,6 +1,6 @@
 package ai.grazie.code.agents.example.planner
 
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.agent.entity.createStorageKey
 import ai.grazie.code.agents.core.agent.entity.LocalAgentNode
