+++ agents/agents-examples/src/main/kotlin/ai/grazie/code/agents/example/structureddata/MarkdownStreamingWithToolsExample.kt
@@ -2,7 +2,7 @@ package ai.grazie.code.agents.example.structureddata
 
 import ai.grazie.code.agents.core.tools.SimpleToolRegistry
 import ai.grazie.code.agents.example.TokenService
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.dsl.builder.forwardTo
 import ai.grazie.code.agents.core.dsl.builder.simpleStrategy
