+++ agents/agents-features/agents-features-memory/src/jvmTest/kotlin/ai/grazie/code/agents/local/memory/MemoryNodesTest.kt
@@ -2,7 +2,7 @@ package ai.grazie.code.agents.local.memory
 
 import ai.grazie.code.agents.core.event.EventHandler
 import ai.grazie.code.agents.core.tools.ToolRegistry
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.dsl.builder.forwardTo
 import ai.grazie.code.agents.core.dsl.builder.strategy
