+++ agents/agents-features/agents-features-trace/src/jvmTest/kotlin/ai/grazie/code/agents/local/features/tracing/writer/TraceTestAPI.kt
@@ -2,7 +2,7 @@ package ai.grazie.code.agents.local.features.tracing.writer
 
 import ai.grazie.code.agents.core.event.EventHandler
 import ai.grazie.code.agents.core.tools.ToolRegistry
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.agent.entity.LocalAgentStrategy
 import ai.jetbrains.code.prompt.dsl.prompt
