+++ prompt/prompt-executor/prompt-executor-llms-all/src/jvmTest/kotlin/ai/jetbrains/code/prompt/executor/llms/all/KotlinAIAgentWithMultipleLLMTest.kt
@@ -1,6 +1,6 @@
 package ai.jetbrains.code.prompt.executor.llms.all
 
-import ai.grazie.code.agents.core.KotlinAIAgent
+import ai.grazie.code.agents.core.agent.KotlinAIAgent
 import ai.grazie.code.agents.core.agent.entity.ContextTransitionPolicy
 import ai.grazie.code.agents.core.agent.config.LocalAgentConfig
 import ai.grazie.code.agents.core.dsl.builder.forwardTo
